Improve the paging API
All clients to know exactly when GET calls might be triggered when
one of the pageable-APIs is used. This is achieved by returning
information about the *current* page on each call to 'next()' of
the underlying generator. The first call might return:
   <object>, 10, 0
indicating this object is the 0th indexed item in the current page
that has 10 items. This means that previous API returned info
about 10 such items. The subsequent call to next() would return:
   <object>, 10, 1
And so forth. Eventually returning:
   <object>, 10, 9

At this point, the next time next() is called a new GET request
will be trigger in order to fetch the next *page* of results. This
might then return:
   <object>, 12, 0
indicating that this next *page* has 12 results.