Add support for As-User and shared items headers.
The Box API supports enterprise admins making requests as users via the As-User header.
It also supports gaining access to items via a shared link via the BoxApi: shared_link= header.

This commit adds an `as_user` and `with_shared_link` method to the SDK client, session, and endpoint objects
to support a fluent interface for adding the appropriate headers to requests.

For example, to get user with id 1234's root folder info:
    client.as_user(User('1234')).folder('0').get()