Rename uses of enterprise_token to enterprise_id
This parameter is referred to as enterprise_id in the Box
Developer documentation. Also, the noun "token" is already used
to refer to the data returned at the end of the auth process. So
rename this to reduce confusion.

This is a backwards compatible change, because the method
parameter name was already correct. This is only a change to
private attributes, test code, and documentation.

Also add a temporary pip version exclusion on the latest version
of pytest, which is broken on Python 2.