Add stream_type param for events long polling
Add a `stream_type` parameter to the
`generate_events_with_long_polling()` method and its helper
methods. This allows users to listen to other event stream
types. The default is still to listen to the 'all' stream.

In order to differentiate between user and enterprise event
streams, the following enum hierarchy was created, with the help
of a new `ExtendableEnumMeta` metaclass:

    +-- `EventsStreamType`
          |
          +-- `UserEventsStreamType`
          |
          +-- `EnterpriseEventsStreamType`

Unrelatedly, fix up the RST for the README. Add a table of
contents, and fix up backticks (double-backticks are needed to
create code spans, single-backticks aren't enough).