Add Event class (#139)
Added an `Event` class, edited `Events.get_event()` to return a list of them.

This creates a new branch in the object class hierarchy, with a distinction between REST objects (instances of `BaseObject` and its subclasses) and non-REST (objects that are not addressable / queryable in the API) objects (instances of `APIJSONObject` and its subclasses). `BaseObject` and `APIJSONObject` inherit from a common baseclass, `BaseAPIJSONObject`, which is now the base for translation.

To maintain backwards-compatibility with methods that used to return `dict`, `APIJSONObject` implements the `Mapping` interface.

Fixes #15.