Refactors to JWTAuth
Rename `JWTAuth.authenticate_app_user()` to
`JWTAuth.authenticate_user()`, to reflect the fact that the new
Box service accounts can authenticate as managed users as well
as app users. `JWTAuth.authenticate_app_user()` is retained as
an alias, for backwards-compatability.

Allow a user ID to be passed to `JWTAuth`, instead of a `User`
object. Previously, developers would need to import and manually
construct a `boxsdk.object.user.User`, since `Client.user()`
cannot be used until the `Client` has been constructed (which is
usually done after all authentication is complete).

Allow a user or user ID to be passed to the `JWTAuth`
constructor. This way, `JWTAuth.authenticate_user()` can be
called without any arguments. More importantly, this means that
`JWTAuth.refresh()` can be called immediately after
construction, with no need for a manual call to
`JWTAuth.authenticate_user()`.

When `auto_session_renewal=True`, if there is no access token,
then `BoxSession.request()` will renew the token _before_ making
the request. This saves an API call. And combined with the
above, authentication for `JWTAuth` objects can be done
completely automatically, at the time of first API call.

Document that the `enterprise_id` argument to `JWTAuth` is
allowed to be `None` (and must be `None` if `user` is passed).
Also allow it to be passed to `authenticate_instance()`, so that
it doesn't need to be provided at construction time.

Clean up the README.

Fixes #174.