Allow passing user and enterprise_id to JWTAuth
The initial commit that added the `user` parameter to `JWTAuth`
only allowed it to be passed when `enterprise_id=None` was
passed.

After playing around with this a little bit, I determined that
this is unnecessarily inflexible for developers. So this
constraint is now relaxed. They can both be passed, with `user`
taking precedence.

Fixup for Pull Request #178.