Pull in repo config changes from master branch
Add testing support for Python 3.6, and upgrade PyPy tests to
use PyPy2.7 5.8.0.

Lock sphinx to a version that still supports Python 2.6. (In the
future, on the master branch, we should exclude the sphinx
install from CI jobs that don't use it, and use Python 3.6 for
the job that does require sphinx.)

Add new pattern to .gitignore, update configs in .pylintrc, add
isort config.