Automatically retry JWT auth API calls that fail because of the exp claim.
JWT auth requires an expiration time, but legitimate authorization
requests can be rejected if the system times differ between the
request originator and the Box servers.

This commit adds logic to detect this situation and automatically retry.
The Box servers include a ``Date`` header when the auth request is
rejected due to the exp claim, so we use that value to construct the
new claim.