Fix tests that were broken on Python 3.7
The test were broken, not the production code.

For test_jwt_auth and test_group, the tests were always subty
wrong, where we'd run out of test data to feed to our mocks
earlier than we were supposed to. But because the StopIteration
was masked, we never noticed. And somehow our assertions didn't
catch these problems.

Now because StopIteration turns into a RuntimeError inside a
generator (PEP 479), these pre-existing bugs were revealed.

For test_events, we were using StopIteration to break out of a
generator. Now we just have to change our test code to use a
different exception.

In test_group, a generator had an incorrect implementation of
Box's paging strategy, which would cause the SDK to try
requesting one extra page of data, which would try to get more
data out of the mock, which would trigger a StopIteration
exception.

In test_jwt_auth, we were only seeding a mock with enough data
for one authentication. But many of our tests do two
authentications. So on old versions of Python, when we'd do the
second authentication, the mock would run out of data, and we'd
seed an empty JTI string. This would silently work before.